add warning about token expiration; small tweaks
