+++ etna/lib/commands.rb
@@ -355,6 +355,8 @@ class EtnaApp
         def execute(project_name, redcap_tokens, models: "all", commit: false)
           raise "Must provide at least one REDCap token (comma-separated)." unless redcap_tokens.split(',').length > 0
 
+          raise "Your token is expired." if polyphemus_client.token_expired?
+
           puts "NOTE: This is a **preview** of what the data loading will look like. Use the --commit flag to load records into Magma." unless commit
 
           polyphemus_client.job(Etna::Clients::Polyphemus::RedcapJobRequest.new(
