+++ etna/lib/etna/clients/janus/client.rb
@@ -2,20 +2,11 @@ require 'net/http/post/multipart'
 require 'singleton'
 require_relative '../../client'
 require_relative './models'
+require_relative '../base_client'
 
 module Etna
   module Clients
-    class Janus
-      def initialize(host:, token:, ignore_ssl: false)
-        raise 'Janus client configuration is missing host.' unless host
-        raise 'Janus client configuration is missing token.' unless token
-        @etna_client = ::Etna::Client.new(
-          host,
-          token,
-          routes_available: false,
-          ignore_ssl: ignore_ssl)
-      end
-
+    class Janus < Etna::Clients::BaseClient
       def get_project(get_project_request = GetProjectRequest.new)
         html = nil
         @etna_client.get(
