+++ etna/lib/etna/clients/polyphemus/client.rb
@@ -2,20 +2,11 @@ require 'net/http/post/multipart'
 require 'singleton'
 require_relative '../../client'
 require_relative './models'
+require_relative '../base_client'
 
 module Etna
   module Clients
-    class Polyphemus
-      def initialize(host:, token:, ignore_ssl: false)
-        raise 'Polyphemus client configuration is missing host.' unless host
-        raise 'Polyphemus client configuration is missing token.' unless token
-        @etna_client = ::Etna::Client.new(
-          host,
-          token,
-          routes_available: false,
-          ignore_ssl: ignore_ssl)
-      end
-
+    class Polyphemus < Etna::Clients::BaseClient
       def configuration(configuration_request = ConfigurationRequest.new)
         json = nil
         @etna_client.get(
@@ -27,7 +18,7 @@ module Etna
         ConfigurationResponse.new(json)
       end
 
-      def job(job_request = JobRequest.new, &block)
+      def job(job_request = JobRequest.new)
         # Because this is a streaming response, just yield the response back.
         #   The consumer will have to iterate over the response.read_body, like
         #
